chore(ui): Removing unused old components, moving new stuff into place.
This removes a ton of the old components that aren't used anymore.
Moving the new components into place. Especially in the `ui/pages`
directory.
