--- ui/components/GlobalFooter.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import { Typography } from '@mui/material';
-
-const GlobalFooter = (): JSX.Element => {
-  return (
-    <Typography
-      className="absolute inline w-full text-center bottom-1 opacity-30 z-50"
-    >
-      © { new Date().getFullYear() } monetr LLC
-    </Typography>
-  );
-};
-
-export default GlobalFooter;
