--- ui/components/Layout/Sidebar/styles/Sidebar.scss
@@ -1,95 +1,0 @@
-@import 'ui/styles/_variables.scss';
-
-div.sidebar {
-  z-index: $sidebar-z-index;
-  background: linear-gradient(to bottom,
-          $primary-color $nav-bar-height, // Start the gradient below the bottom of the navbar.
-          $primary-color, darken($primary-color, 7));
-  scroll-behavior: auto;
-  overflow-y: auto;
-
-  button.MuiButton-root, a.MuiButton-root {
-    text-transform: unset;
-  }
-
-  @media (min-width: 720px) {
-    div.sidebar-button-wrapper {
-      margin-bottom: -5px;
-      margin-top: -5px;
-
-      &:not(.sidebar-button-wrapper-active) {
-        height: 60px;
-        padding-top: 10px;
-        padding-bottom: 10px;
-
-        button.MuiButton-root, a.MuiButton-root {
-          border-top-right-radius: 0 !important;
-          border-bottom-right-radius: 0 !important;
-
-          > span.MuiTouchRipple-root {
-            border-top-right-radius: inherit;
-            border-bottom-right-radius: inherit;
-            background-color: inherit;
-            transition: opacity 500ms;
-          }
-        }
-      }
-
-      &.sidebar-button-wrapper-active {
-        div.navigation-before {
-          width: 100%;
-          position: relative;
-          right: 0;
-          top: 0;
-          height: 10px;
-          background: white;
-
-          &::before {
-            display: block;
-            content: " ";
-            color: white;
-            height: 10px;
-            width: 100%;
-            background: $primary-color;
-            border-radius: 0 0 10px 0;
-          }
-        }
-
-        div.navigation-after {
-          width: 100%;
-          position: relative;
-          right: 0;
-          bottom: 0;
-          height: 10px;
-          background: white;
-
-          &::before {
-            display: block;
-            content: " ";
-            color: white;
-            height: 10px;
-            width: 100%;
-            background: $primary-color;
-            border-radius: 0 10px 0 0;
-          }
-        }
-
-        button.MuiButton-root, a.MuiButton-root {
-          border-top-right-radius: 0 !important;
-          border-bottom-right-radius: 0 !important;
-          color: $primary-color;
-          background: linear-gradient(to right,
-                  lighten($primary-color, 50) 50px, // Start the gradient below the bottom of the navbar.
-                  white);
-
-          > span.MuiTouchRipple-root {
-            opacity: 0;
-            border-top-right-radius: inherit;
-            border-bottom-right-radius: inherit;
-            background-color: inherit;
-          }
-        }
-      }
-    }
-  }
-}
