+++ ui/pages/password/reset.tsx
@@ -1,13 +1,15 @@
-import React, { Fragment, useEffect } from 'react';
+import React, { useEffect } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
-import { Button, CircularProgress, TextField } from '@mui/material';
-import { Formik, FormikErrors, FormikHelpers } from 'formik';
+import { FormikErrors, FormikHelpers } from 'formik';
 import { useSnackbar } from 'notistack';
 
-import BackToLoginButton from 'components/Authentication/BackToLoginButton';
-import CenteredLogo from 'components/Logo/CenteredLogo';
+import MFormButton from 'components/MButton';
+import MForm from 'components/MForm';
+import MLink from 'components/MLink';
+import MLogo from 'components/MLogo';
+import MSpan from 'components/MSpan';
+import MTextField from 'components/MTextField';
 import useResetPassword from 'hooks/useResetPassword';
-import mergeTailwind from 'util/mergeTailwind';
 
 interface ResetPasswordValues {
   password: string;
@@ -19,14 +21,13 @@ const initialValues: ResetPasswordValues = {
   verifyPassword: '',
 };
 
-export default function ResetPasswordPage(): JSX.Element {
+export default function PasswordResetNew(): JSX.Element {
   const { enqueueSnackbar } = useSnackbar();
   const location = useLocation();
   const navigate = useNavigate();
   const resetPassword = useResetPassword();
-
   const { state: routeState } = useLocation();
-
+  const message = (routeState && routeState['message']) || 'Enter the new password you would like to use.';
   const search = location.search;
   const query = new URLSearchParams(search);
   // The token is loaded from the route state (which is provided when a password reset is being forced) or from the
@@ -35,6 +36,7 @@ export default function ResetPasswordPage(): JSX.Element {
 
   useEffect(() => {
     if (!token) {
+      // TODO This ends up enqueuing two snackbars for some reason?
       navigate('/login');
       enqueueSnackbar('You must get a password reset link to change your password.', {
         variant: 'warning',
@@ -45,9 +47,19 @@ export default function ResetPasswordPage(): JSX.Element {
     // Clear the URL so that the token is not shown. But also so that the user cannot accidentally navigate back to the
     // password reset page with the token still in place.
     window.history.replaceState({}, document.title, !token ? '/login' : location.pathname);
-  }, [routeState, token, enqueueSnackbar, navigate, location.pathname]);
+  }, [token, enqueueSnackbar, navigate, location.pathname]);
 
-  function validateInput(values: ResetPasswordValues): FormikErrors<ResetPasswordValues> {
+  async function submit(values: ResetPasswordValues, helpers: FormikHelpers<ResetPasswordValues>): Promise<void> {
+    helpers.setSubmitting(true);
+
+    return resetPassword(values.password, token)
+      // If the reset password fails, then set submitting to false and do nothing. The error will already have been
+      // displayed by the reset password function. We only do this if there is an error because if this succeeds then
+      // the user is automatically redirected to the login page.
+      .catch(() => helpers.setSubmitting(false));
+  }
+
+  function validate(values: ResetPasswordValues): FormikErrors<ResetPasswordValues> {
     const errors: FormikErrors<ResetPasswordValues> = {};
 
     if (values.password) {
@@ -63,106 +75,51 @@ export default function ResetPasswordPage(): JSX.Element {
     return errors;
   }
 
-  async function submitResetPassword(values: ResetPasswordValues, helpers: FormikHelpers<ResetPasswordValues>): Promise<void> {
-    helpers.setSubmitting(true);
-
-    return resetPassword(values.password, token)
-      // If the reset password fails, then set submitting to false and do nothing. The error will already have been
-      // displayed by the reset password function. We only do this if there is an error because if this succeeds then
-      // the user is automatically redirected to the login page.
-      .catch(() => helpers.setSubmitting(false));
-  }
-
-  const message = (routeState && routeState['message']) || 'Enter the new password you would like to use.';
-
   return (
-    <Fragment>
-      <BackToLoginButton />
-      <Formik
-        initialValues={ initialValues }
-        validate={ validateInput }
-        onSubmit={ submitResetPassword }
-      >
-        { ({
-          values,
-          errors,
-          touched,
-          handleChange,
-          handleBlur,
-          handleSubmit,
-          isSubmitting,
-          submitForm,
-        }) => (
-          <form onSubmit={ handleSubmit } className="h-full overflow-y-auto pb-20">
-            <div className="flex items-center justify-center w-full h-full max-h-full">
-              <div className="w-full p-10 xl:w-3/12 lg:w-5/12 md:w-2/3 sm:w-10/12 max-w-screen-sm sm:p-0">
-                <CenteredLogo />
-                <div className="w-full">
-                  <div className="w-full pb-2.5">
-                    <p className="text-center">
-                      { message }
-                    </p>
-                  </div>
-                  <div className="w-full pb-2.5">
-                    <TextField
-                      autoComplete="current-password"
-                      autoFocus
-                      className="w-full"
-                      disabled={ isSubmitting }
-                      error={ touched.password && !!errors.password }
-                      helperText={ (touched.password && errors.password) ? errors.password : null }
-                      id="reset-password"
-                      label="Password"
-                      name="password"
-                      onBlur={ handleBlur }
-                      onChange={ handleChange }
-                      value={ values.password }
-                      variant="outlined"
-                      type="password"
-                    />
-                  </div>
-                  <div className="w-full pb-2.5">
-                    <TextField
-                      autoComplete="current-password"
-                      className="w-full"
-                      disabled={ isSubmitting }
-                      error={ touched.verifyPassword && !!errors.verifyPassword }
-                      helperText={ (touched.verifyPassword && errors.verifyPassword) ? errors.verifyPassword : null }
-                      id="reset-password-verify"
-                      label="Verify Password"
-                      name="verifyPassword"
-                      onBlur={ handleBlur }
-                      onChange={ handleChange }
-                      value={ values.verifyPassword }
-                      variant="outlined"
-                      type="password"
-                    />
-                  </div>
-                  <div className="w-full pt-2.5 mb-10">
-                    <Button
-                      className="w-full"
-                      color="primary"
-                      disabled={ isSubmitting || !values.password || !values.verifyPassword }
-                      onClick={ submitForm }
-                      type="submit"
-                      variant="contained"
-                    >
-                      { isSubmitting && <CircularProgress
-                        className={ mergeTailwind('mr-2', {
-                          'opacity-50': isSubmitting,
-                        }) }
-                        size="1em"
-                        thickness={ 5 }
-                      /> }
-                      { isSubmitting ? 'Resetting Password...' : 'Reset Password' }
-                    </Button>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </form>
-        ) }
-      </Formik>
-    </Fragment>
+    <MForm
+      onSubmit={ submit }
+      initialValues={ initialValues }
+      validate={ validate }
+      className="w-full h-full flex flex-col pt-10 md:pt-0 mb:pb-10 md:justify-center items-center px-5 gap-1"
+    >
+      <div className='flex items-center flex-col gap-1 w-full xl:w-1/5 lg:w-1/4 md:w-1/3 sm:w-1/2'>
+        <div className="max-w-[128px] w-full">
+          <MLogo />
+        </div>
+        <MSpan className='flex items-center text-center'>
+          { message }
+        </MSpan>
+        <MTextField
+          autoFocus
+          autoComplete='current-password'
+          label="Password"
+          name='password'
+          type='password'
+          required
+          className="w-full"
+        />
+        <MTextField
+          autoComplete='current-password'
+          label="Verify Password"
+          name='verifyPassword'
+          type='password'
+          required
+          className="w-full"
+        />
+        <MFormButton
+          color="primary"
+          variant="solid"
+          role="form"
+          type="submit"
+          className='w-full'
+        >
+          Reset Password
+        </MFormButton>
+      </div>
+      <div className="w-full lg:w-1/4 sm:w-1/3 mt-1 flex justify-center gap-1">
+        <MSpan color="subtle" className='text-sm'>Remembered your password?</MSpan>
+        <MLink to="/login" size="sm">Sign in</MLink>
+      </div>
+    </MForm>
   );
 }
