--- ui/pages/transactions-new.stories.tsx
@@ -1,705 +1,0 @@
-import React from 'react';
-import { Meta, StoryObj } from '@storybook/react';
-
-import TransactionsNew from './transactions-new';
-
-import MLayout from 'components/MLayout';
-import { rest } from 'msw';
-
-export default {
-  title: 'Pages/Transactions',
-  component: TransactionsNew,
-} as Meta<typeof TransactionsNew>;
-
-export const Default: StoryObj = {
-  name: 'Default',
-  render: () => (
-    <MLayout>
-      <TransactionsNew />
-    </MLayout>
-  ),
-  parameters: {
-    msw: {
-      handlers: [
-        rest.get('/api/config', (_req, res, ctx) => {
-          return res(ctx.json({}));
-        }),
-        rest.get('/api/links', (_req, res, ctx) => {
-          return res(ctx.json([
-            {
-              'linkId': 4,
-              'linkType': 1,
-              'plaidInstitutionId': 'ins_116794',
-              'plaidNewAccountsAvailable': false,
-              'linkStatus': 2,
-              'expirationDate': null,
-              'institutionName': 'Mercury',
-              'description': null,
-              'createdAt': '2022-09-25T02:08:40.758642Z',
-              'createdByUserId': 1,
-              'updatedAt': '2023-05-13T09:01:24.952957Z',
-              'lastManualSync': '2023-05-02T19:56:34.953077Z',
-              'lastSuccessfulUpdate': '2023-05-13T09:01:24.952559Z',
-            },
-          ]));
-        }),
-        rest.get('/api/bank_accounts', (_req, res, ctx) => {
-          return res(ctx.json([
-            {
-              'bankAccountId': 1,
-              'linkId': 4,
-              'availableBalance': 70479,
-              'currentBalance': 70479,
-              'mask': '1234',
-              'name': 'Mercury Checking',
-              'originalName': 'Mercury Checking',
-              'officialName': 'Mercury Checking',
-              'accountType': 'depository',
-              'accountSubType': 'checking',
-              'status': 'active',
-              'lastUpdated': '2023-05-13T09:01:24.147095Z',
-            },
-          ]));
-        }),
-        rest.get('/api/bank_accounts/1/balances', (_req, res, ctx) => {
-          return res(ctx.json({
-            'bankAccountId': 1,
-            'current': 70479,
-            'available': 70479,
-            'safe': 25012,
-            'expenses': 13867,
-            'goals': 31600,
-          }));
-        }),
-        rest.get('/api/bank_accounts/1/spending', (_req, res, ctx) => {
-          return res(ctx.json([
-            {
-              'spendingId': 136,
-              'bankAccountId': 1,
-              'fundingScheduleId': 3,
-              'spendingType': 0,
-              'name': 'BuildKite',
-              'description': 'Every month on the 27th',
-              'targetAmount': 1500,
-              'currentAmount': 0,
-              'usedAmount': 0,
-              'recurrenceRule': 'FREQ=MONTHLY;INTERVAL=1;BYMONTHDAY=27',
-              'lastRecurrence': '2023-05-27T05:00:00Z',
-              'nextRecurrence': '2023-05-27T05:00:00Z',
-              'nextContributionAmount': 1500,
-              'isBehind': false,
-              'isPaused': false,
-              'dateCreated': '2021-12-14T16:42:11Z',
-              'dateStarted': '2023-02-28T06:00:00Z',
-            }, {
-              'spendingId': 137,
-              'bankAccountId': 1,
-              'fundingScheduleId': 3,
-              'spendingType': 0,
-              'name': 'Sentry',
-              'description': 'Every month on the 25th',
-              'targetAmount': 2900,
-              'currentAmount': 0,
-              'usedAmount': 0,
-              'recurrenceRule': 'FREQ=MONTHLY;INTERVAL=1;BYMONTHDAY=25',
-              'lastRecurrence': '2023-05-25T05:00:00Z',
-              'nextRecurrence': '2023-05-25T05:00:00Z',
-              'nextContributionAmount': 2900,
-              'isBehind': false,
-              'isPaused': false,
-              'dateCreated': '2021-12-14T16:42:41Z',
-              'dateStarted': '2023-03-25T05:00:00Z',
-            }, {
-              'spendingId': 201,
-              'bankAccountId': 1,
-              'fundingScheduleId': 3,
-              'spendingType': 1,
-              'name': 'Rainy Day',
-              'targetAmount': 10000,
-              'currentAmount': 31600,
-              'usedAmount': 0,
-              'recurrenceRule': null,
-              'lastRecurrence': '2022-12-31T06:00:00Z',
-              'nextRecurrence': '2022-12-31T06:00:00Z',
-              'nextContributionAmount': 0,
-              'isBehind': false,
-              'isPaused': false,
-              'dateCreated': '2022-11-29T16:32:58Z',
-              'dateStarted': '2022-12-31T06:00:00Z',
-            }, {
-              'spendingId': 192,
-              'bankAccountId': 1,
-              'fundingScheduleId': 3,
-              'spendingType': 0,
-              'name': 'Cloud Staging',
-              'description': 'Every month on the 1st',
-              'targetAmount': 10000,
-              'currentAmount': 1874,
-              'usedAmount': 0,
-              'recurrenceRule': 'FREQ=MONTHLY;INTERVAL=1;BYMONTHDAY=1',
-              'lastRecurrence': '2023-06-01T05:00:00Z',
-              'nextRecurrence': '2023-06-01T05:00:00Z',
-              'nextContributionAmount': 4063,
-              'isBehind': false,
-              'isPaused': false,
-              'dateCreated': '2022-11-07T15:09:32Z',
-              'dateStarted': '2023-03-01T06:00:00Z',
-            }, {
-              'spendingId': 189,
-              'bankAccountId': 1,
-              'fundingScheduleId': 3,
-              'spendingType': 0,
-              'name': 'Google Voice',
-              'description': 'Every month on the 1st',
-              'targetAmount': 1366,
-              'currentAmount': 21,
-              'usedAmount': 0,
-              'recurrenceRule': 'FREQ=MONTHLY;INTERVAL=1;BYMONTHDAY=1',
-              'lastRecurrence': '2023-06-01T05:00:00Z',
-              'nextRecurrence': '2023-06-01T05:00:00Z',
-              'nextContributionAmount': 672,
-              'isBehind': false,
-              'isPaused': false,
-              'dateCreated': '2022-11-02T14:11:24Z',
-              'dateStarted': '2023-03-01T06:00:00Z',
-            }, {
-              'spendingId': 208,
-              'bankAccountId': 1,
-              'fundingScheduleId': 3,
-              'spendingType': 0,
-              'name': 'Google Domains ($12)',
-              'description': 'Every year on the 29th of January',
-              'targetAmount': 1200,
-              'currentAmount': 347,
-              'usedAmount': 0,
-              'recurrenceRule': 'FREQ=YEARLY;INTERVAL=1;BYMONTH=1;BYMONTHDAY=29',
-              'lastRecurrence': null,
-              'nextRecurrence': '2024-01-29T06:00:00Z',
-              'nextContributionAmount': 50,
-              'isBehind': false,
-              'isPaused': false,
-              'dateCreated': '2023-01-30T22:16:20Z',
-              'dateStarted': '2024-01-29T06:00:00Z',
-            }, {
-              'spendingId': 171,
-              'bankAccountId': 1,
-              'fundingScheduleId': 3,
-              'spendingType': 0,
-              'name': 'ngrok',
-              'description': 'Every year on the 26th of June',
-              'targetAmount': 6000,
-              'currentAmount': 5184,
-              'usedAmount': 0,
-              'recurrenceRule': 'FREQ=YEARLY;INTERVAL=1;BYMONTH=6;BYMONTHDAY=26',
-              'lastRecurrence': null,
-              'nextRecurrence': '2023-06-25T05:00:00Z',
-              'nextContributionAmount': 204,
-              'isBehind': false,
-              'isPaused': false,
-              'dateCreated': '2022-06-28T15:59:10Z',
-              'dateStarted': '2023-06-25T05:00:00Z',
-            }, {
-              'spendingId': 134,
-              'bankAccountId': 1,
-              'fundingScheduleId': 3,
-              'spendingType': 0,
-              'name': 'Freshbooks',
-              'description': 'Every month on the 10th',
-              'targetAmount': 1700,
-              'currentAmount': 0,
-              'usedAmount': 0,
-              'recurrenceRule': 'FREQ=MONTHLY;INTERVAL=1;BYMONTHDAY=10',
-              'lastRecurrence': '2023-06-10T05:00:00Z',
-              'nextRecurrence': '2023-06-10T05:00:00Z',
-              'nextContributionAmount': 850,
-              'isBehind': false,
-              'isPaused': false,
-              'dateCreated': '2021-12-14T16:40:46Z',
-              'dateStarted': '2023-03-10T06:00:00Z',
-            }, {
-              'spendingId': 138,
-              'bankAccountId': 1,
-              'fundingScheduleId': 3,
-              'spendingType': 0,
-              'name': 'GitHub',
-              'description': 'Every month on the 19th',
-              'targetAmount': 2600,
-              'currentAmount': 2600,
-              'usedAmount': 0,
-              'recurrenceRule': 'FREQ=MONTHLY;INTERVAL=1;BYMONTHDAY=19',
-              'lastRecurrence': '2023-05-19T05:00:00Z',
-              'nextRecurrence': '2023-05-19T05:00:00Z',
-              'nextContributionAmount': 0,
-              'isBehind': false,
-              'isPaused': false,
-              'dateCreated': '2021-12-14T16:43:04Z',
-              'dateStarted': '2023-03-19T05:00:00Z',
-            }, {
-              'spendingId': 191,
-              'bankAccountId': 1,
-              'fundingScheduleId': 3,
-              'spendingType': 0,
-              'name': 'Cloud Production',
-              'description': 'Every month on the 1st',
-              'targetAmount': 28000,
-              'currentAmount': 3841,
-              'usedAmount': 0,
-              'recurrenceRule': 'FREQ=MONTHLY;INTERVAL=1;BYMONTHDAY=1',
-              'lastRecurrence': '2023-06-01T05:00:00Z',
-              'nextRecurrence': '2023-06-01T05:00:00Z',
-              'nextContributionAmount': 12079,
-              'isBehind': false,
-              'isPaused': false,
-              'dateCreated': '2022-11-07T15:09:16Z',
-              'dateStarted': '2023-03-01T06:00:00Z',
-            }, {
-              'spendingId': 135,
-              'bankAccountId': 1,
-              'fundingScheduleId': 3,
-              'spendingType': 0,
-              'name': 'G-Suite ($12)',
-              'description': 'Every month on the 1st',
-              'targetAmount': 1200,
-              'currentAmount': 0,
-              'usedAmount': 0,
-              'recurrenceRule': 'FREQ=MONTHLY;INTERVAL=1;BYMONTHDAY=1',
-              'lastRecurrence': '2023-06-01T05:00:00Z',
-              'nextRecurrence': '2023-06-01T05:00:00Z',
-              'nextContributionAmount': 600,
-              'isBehind': false,
-              'isPaused': false,
-              'dateCreated': '2021-12-14T16:41:18Z',
-              'dateStarted': '2023-03-01T06:00:00Z',
-            },
-          ]));
-        }),
-        rest.get('/api/bank_accounts/1/transactions', (_req, res, ctx) => {
-          return res(ctx.json([
-            {
-              'transactionId': 10657,
-              'bankAccountId': 1,
-              'amount': -25000,
-              'spendingId': null,
-              'categories': ['Transfer', 'Payroll'],
-              'originalCategories': ['Transfer', 'Payroll'],
-              'date': '2023-05-13T05:00:00Z',
-              'authorizedDate': null,
-              'name': '1234 TREASURY PR; 1234 TREASURY PR:1234567:COURANT ELLIOT; 123456. Merchant name: 1234 TREASURY PR',
-              'customName': null,
-              'originalName': '1234 TREASURY PR; 1234 TREASURY PR:1234567:COURANT ELLIOT; 123456. Merchant name: 1234 TREASURY PR',
-              'originalMerchantName': '',
-              'isPending': true,
-              'createdAt': '2023-05-13T09:01:24.147095Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 10620,
-              'bankAccountId': 1,
-              'amount': 1700,
-              'spendingId': 134,
-              'spendingAmount': 1700,
-              'categories': ['Service', 'Financial', 'Accounting and Bookkeeping'],
-              'originalCategories': ['Service', 'Financial', 'Accounting and Bookkeeping'],
-              'date': '2023-05-10T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'FreshBooks',
-              'customName': null,
-              'originalName': 'FRESHBOOKS. Merchant name: FRESHBOOKS',
-              'merchantName': 'FreshBooks',
-              'originalMerchantName': 'FreshBooks',
-              'isPending': false,
-              'createdAt': '2023-05-10T06:28:37Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 10576,
-              'bankAccountId': 1,
-              'amount': 24159,
-              'spendingId': 191,
-              'spendingAmount': 24159,
-              'categories': ['Shops', 'Digital Purchase'],
-              'originalCategories': ['Shops', 'Digital Purchase'],
-              'date': '2023-05-05T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'Monetr LLC',
-              'customName': null,
-              'originalName': 'GOOGLE; GOOGLE:013190208 DEBIT:monetr LLC; US003UFE4R. Merchant name: GOOGLE',
-              'merchantName': 'Monetr LLC',
-              'originalMerchantName': 'Monetr LLC',
-              'isPending': false,
-              'createdAt': '2023-05-07T03:48:38Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 10546,
-              'bankAccountId': 1,
-              'amount': 8126,
-              'spendingId': 192,
-              'spendingAmount': 8126,
-              'categories': ['Shops', 'Digital Purchase'],
-              'originalCategories': ['Shops', 'Digital Purchase'],
-              'date': '2023-05-05T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'Monetr LLC',
-              'customName': null,
-              'originalName': 'GOOGLE; GOOGLE:013190208 DEBIT:monetr LLC; US003UGB8G. Merchant name: GOOGLE',
-              'merchantName': 'Monetr LLC',
-              'originalMerchantName': 'Monetr LLC',
-              'isPending': false,
-              'createdAt': '2023-05-06T02:24:16Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 10483,
-              'bankAccountId': 1,
-              'amount': 1345,
-              'spendingId': 189,
-              'spendingAmount': 1345,
-              'categories': ['Shops', 'Digital Purchase'],
-              'originalCategories': ['Shops', 'Digital Purchase'],
-              'date': '2023-05-02T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'Svcsmonetr.app.',
-              'customName': null,
-              'originalName': 'SVCSmonetr.app. Merchant name: Google Pay',
-              'merchantName': 'Svcsmonetr.app.',
-              'originalMerchantName': 'Svcsmonetr.app.',
-              'isPending': false,
-              'createdAt': '2023-05-02T04:48:33Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 10476,
-              'bankAccountId': 1,
-              'amount': 1200,
-              'spendingId': 135,
-              'spendingAmount': 1200,
-              'categories': ['Shops', 'Digital Purchase'],
-              'originalCategories': ['Shops', 'Digital Purchase'],
-              'date': '2023-05-01T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'Google Workspace',
-              'customName': null,
-              'originalName': 'GSUITE_monetr.app. Merchant name: Google Pay',
-              'merchantName': 'Google Workspace',
-              'originalMerchantName': 'Google Workspace',
-              'isPending': false,
-              'createdAt': '2023-05-01T16:28:04Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 10426,
-              'bankAccountId': 1,
-              'amount': 1500,
-              'spendingId': 136,
-              'spendingAmount': 1500,
-              'categories': ['Transfer', 'Debit'],
-              'originalCategories': ['Transfer', 'Debit'],
-              'date': '2023-04-28T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'Buildkite.',
-              'customName': null,
-              'originalName': 'BUILDKITE. Merchant name: BUILDKITE',
-              'merchantName': 'Buildkite.',
-              'originalMerchantName': 'Buildkite.',
-              'isPending': false,
-              'createdAt': '2023-04-28T06:37:50Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 10425,
-              'bankAccountId': 1,
-              'amount': -25000,
-              'spendingId': null,
-              'categories': ['Transfer', 'Payroll'],
-              'originalCategories': ['Transfer', 'Payroll'],
-              'date': '2023-04-28T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'Elliot\'s Contribution',
-              'customName': null,
-              'originalName': '1234 TREASURY PR; 1234 TREASURY PR:1234567:COURANT ELLIOT; 123456. Merchant name: 1234 TREASURY PR',
-              'originalMerchantName': '',
-              'isPending': false,
-              'createdAt': '2023-04-28T06:37:50Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 10404,
-              'bankAccountId': 1,
-              'amount': 2900,
-              'spendingId': 137,
-              'spendingAmount': 2900,
-              'categories': ['Shops', 'Digital Purchase'],
-              'originalCategories': ['Shops', 'Digital Purchase'],
-              'date': '2023-04-27T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'Sentry',
-              'customName': null,
-              'originalName': 'SENTRY. Merchant name: SENTRY',
-              'merchantName': 'Sentry',
-              'originalMerchantName': 'Sentry',
-              'isPending': false,
-              'createdAt': '2023-04-27T12:03:36Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 10291,
-              'bankAccountId': 1,
-              'amount': 1223,
-              'spendingId': 138,
-              'spendingAmount': 1223,
-              'categories': ['Service', 'Computers', 'Software Development'],
-              'originalCategories': ['Service', 'Computers', 'Software Development'],
-              'date': '2023-04-19T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'GitHub',
-              'customName': null,
-              'originalName': 'GITHUB, INC.. Merchant name: GITHUB, INC.',
-              'merchantName': 'GitHub',
-              'originalMerchantName': 'GitHub',
-              'isPending': false,
-              'createdAt': '2023-04-19T05:29:21Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 10290,
-              'bankAccountId': 1,
-              'amount': 1000,
-              'spendingId': 138,
-              'spendingAmount': 1000,
-              'categories': ['Service', 'Computers', 'Software Development'],
-              'originalCategories': ['Service', 'Computers', 'Software Development'],
-              'date': '2023-04-19T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'GitHub',
-              'customName': null,
-              'originalName': 'GITHUB. Merchant name: GITHUB',
-              'merchantName': 'GitHub',
-              'originalMerchantName': 'GitHub',
-              'isPending': false,
-              'createdAt': '2023-04-19T05:29:21Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 10257,
-              'bankAccountId': 1,
-              'amount': 13800,
-              'spendingId': null,
-              'categories': ['Shops', 'Digital Purchase'],
-              'originalCategories': ['Shops', 'Digital Purchase'],
-              'date': '2023-04-17T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'Domains.com',
-              'customName': null,
-              'originalName': 'DOMAINS. Merchant name: Google Pay',
-              'merchantName': 'Domains.com',
-              'originalMerchantName': 'Domains.com',
-              'isPending': false,
-              'createdAt': '2023-04-17T03:45:51.510718Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 10171,
-              'bankAccountId': 1,
-              'amount': -25000,
-              'spendingId': null,
-              'categories': ['Transfer', 'Payroll'],
-              'originalCategories': ['Transfer', 'Payroll'],
-              'date': '2023-04-14T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'Elliot\'s Contribution',
-              'customName': null,
-              'originalName': '1234 TREASURY PR; 1234 TREASURY PR:1234567:COURANT ELLIOT; 123456. Merchant name: 1234 TREASURY PR',
-              'originalMerchantName': '',
-              'isPending': false,
-              'createdAt': '2023-04-14T06:22:12Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 10106,
-              'bankAccountId': 1,
-              'amount': 1700,
-              'spendingId': 134,
-              'spendingAmount': 1500,
-              'categories': ['Service', 'Financial', 'Accounting and Bookkeeping'],
-              'originalCategories': ['Service', 'Financial', 'Accounting and Bookkeeping'],
-              'date': '2023-04-09T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'FreshBooks',
-              'customName': null,
-              'originalName': 'FRESHBOOKS. Merchant name: FRESHBOOKS',
-              'merchantName': 'FreshBooks',
-              'originalMerchantName': 'FreshBooks',
-              'isPending': false,
-              'createdAt': '2023-04-09T14:08:16Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 10040,
-              'bankAccountId': 1,
-              'amount': 24754,
-              'spendingId': 191,
-              'spendingAmount': 24754,
-              'categories': ['Shops', 'Digital Purchase'],
-              'originalCategories': ['Shops', 'Digital Purchase'],
-              'date': '2023-04-05T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'Monetr LLC',
-              'customName': null,
-              'originalName': 'GOOGLE; GOOGLE:013153739 DEBIT:monetr LLC; US003TYTY4. Merchant name: GOOGLE',
-              'merchantName': 'Monetr LLC',
-              'originalMerchantName': 'Monetr LLC',
-              'isPending': false,
-              'createdAt': '2023-04-05T23:14:18Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 10039,
-              'bankAccountId': 1,
-              'amount': 8363,
-              'spendingId': 192,
-              'spendingAmount': 8363,
-              'categories': ['Shops', 'Digital Purchase'],
-              'originalCategories': ['Shops', 'Digital Purchase'],
-              'date': '2023-04-05T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'Monetr LLC',
-              'customName': null,
-              'originalName': 'GOOGLE; GOOGLE:013148094 DEBIT:monetr LLC; US003TYTU4. Merchant name: GOOGLE',
-              'merchantName': 'Monetr LLC',
-              'originalMerchantName': 'Monetr LLC',
-              'isPending': false,
-              'createdAt': '2023-04-05T23:14:18Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 9957,
-              'bankAccountId': 1,
-              'amount': 1200,
-              'spendingId': 135,
-              'spendingAmount': 1200,
-              'categories': ['Shops', 'Digital Purchase'],
-              'originalCategories': ['Shops', 'Digital Purchase'],
-              'date': '2023-04-02T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'Google Workspace',
-              'customName': null,
-              'originalName': 'GSUITE MONETR.A. Merchant name: Google Pay',
-              'merchantName': 'Google Workspace',
-              'originalMerchantName': 'Google Workspace',
-              'isPending': false,
-              'createdAt': '2023-04-02T13:17:53Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 9955,
-              'bankAccountId': 1,
-              'amount': 1366,
-              'spendingId': 189,
-              'spendingAmount': 1344,
-              'categories': ['Shops', 'Digital Purchase'],
-              'originalCategories': ['Shops', 'Digital Purchase'],
-              'date': '2023-04-02T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'Svcsmonetr.app.',
-              'customName': null,
-              'originalName': 'SVCSmonetr.app. Merchant name: Google Pay',
-              'merchantName': 'Svcsmonetr.app.',
-              'originalMerchantName': 'Svcsmonetr.app.',
-              'isPending': false,
-              'createdAt': '2023-04-02T06:57:29Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 9926,
-              'bankAccountId': 1,
-              'amount': -25000,
-              'spendingId': null,
-              'categories': ['Transfer', 'Payroll'],
-              'originalCategories': ['Transfer', 'Payroll'],
-              'date': '2023-03-31T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'Elliot\'s Contribution',
-              'customName': null,
-              'originalName': '1234 TREASURY PR; 1234 TREASURY PR:1234567:COURANT ELLIOT; 123456. Merchant name: 1234 TREASURY PR',
-              'originalMerchantName': '',
-              'isPending': false,
-              'createdAt': '2023-03-31T11:30:38Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 9855,
-              'bankAccountId': 1,
-              'amount': 1500,
-              'spendingId': 136,
-              'spendingAmount': 1500,
-              'categories': ['Transfer', 'Debit'],
-              'originalCategories': ['Transfer', 'Debit'],
-              'date': '2023-03-28T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'Buildkite.',
-              'customName': null,
-              'originalName': 'BUILDKITE. Merchant name: BUILDKITE',
-              'merchantName': 'Buildkite.',
-              'originalMerchantName': 'Buildkite.',
-              'isPending': false,
-              'createdAt': '2023-03-28T01:59:36Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 9848,
-              'bankAccountId': 1,
-              'amount': 2900,
-              'spendingId': 137,
-              'spendingAmount': 2900,
-              'categories': ['Shops', 'Digital Purchase'],
-              'originalCategories': ['Shops', 'Digital Purchase'],
-              'date': '2023-03-27T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'Sentry',
-              'customName': null,
-              'originalName': 'SENTRY. Merchant name: SENTRY',
-              'merchantName': 'Sentry',
-              'originalMerchantName': 'Sentry',
-              'isPending': false,
-              'createdAt': '2023-03-27T13:30:01Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 9753,
-              'bankAccountId': 1,
-              'amount': 1200,
-              'spendingId': 138,
-              'spendingAmount': 1200,
-              'categories': ['Service', 'Computers', 'Software Development'],
-              'originalCategories': ['Service', 'Computers', 'Software Development'],
-              'date': '2023-03-19T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'GitHub',
-              'customName': null,
-              'originalName': 'GITHUB, INC.. Merchant name: GITHUB, INC.',
-              'merchantName': 'GitHub',
-              'originalMerchantName': 'GitHub',
-              'isPending': false,
-              'createdAt': '2023-03-19T06:36:26Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 9752,
-              'bankAccountId': 1,
-              'amount': 1000,
-              'spendingId': 138,
-              'spendingAmount': 1000,
-              'categories': ['Service', 'Computers', 'Software Development'],
-              'originalCategories': ['Service', 'Computers', 'Software Development'],
-              'date': '2023-03-19T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'GitHub',
-              'customName': null,
-              'originalName': 'GITHUB. Merchant name: GITHUB',
-              'merchantName': 'GitHub',
-              'originalMerchantName': 'GitHub',
-              'isPending': false,
-              'createdAt': '2023-03-19T06:36:26Z',
-              'deletedAt': null,
-            }, {
-              'transactionId': 9675,
-              'bankAccountId': 1,
-              'amount': -25000,
-              'spendingId': null,
-              'categories': ['Transfer', 'Payroll'],
-              'originalCategories': ['Transfer', 'Payroll'],
-              'date': '2023-03-14T05:00:00Z',
-              'authorizedDate': null,
-              'name': 'Elliot\'s Contribution',
-              'customName': null,
-              'originalName': 'BAMBOOHR PAYROLL; BAMBOOHR PAYROLL::ELLIOT COURANT; 123456. Merchant name: BAMBOOHR PAYROLL',
-              'originalMerchantName': '',
-              'isPending': false,
-              'createdAt': '2023-03-14T07:07:49Z',
-              'deletedAt': null,
-            },
-          ]));
-        }),
-      ],
-    },
-  },
-};
